==> Pizza/Pizza_code/14_target.py <==
# review data target analysis

'''
[v] 브랜드별 별점 평균 순위 확인
[v] 주문그룹별 주문건수 파악 및 주문그룹별 선호 메뉴구분 분석
[v] 서울권/비서울권 브랜드 별점 합계 & 리뷰 건수 파악을 통한 브랜드 선호도 파악
[v] 브랜드별 메뉴구분별 대표 메뉴 점수 분석 (값은 별점 합계로 계산)
[v] 메뉴구분별 별점 만족도 분석(1~2점 불만족, 3점 보통, 4~5점 만족)
[X] 브랜드 메뉴구분별 리뷰 키워드 top10 분석(기타 제외) Konlpy 문제 해결 후 진행
'''

import pandas as pd
import numpy as np
import os
from openpyxl import Workbook #  결과물을 엑셀 파일로 저장

os.chdir(r"C:\Users\vandr\OneDrive\바탕 화면\Bigdata\Project_python\Pizza\Dataset\2차 분석")
# df = pd.read_excel("Review.xlsx") # 전체 리뷰
df = pd.read_excel("sample.xlsx") # 샘플 테스트
book = pd.read_excel("MENU.xlsx")

'''
# 전체 데이터 불러온 뒤 샘플데이터 생성
sp = df.sample(frac = 0.2, replace = True)
sp.to_excel("sample.xlsx")
'''

'''
# 브랜드명 데이터 개수 확인 및 별점 평균 순위 확인
df3 = df.groupby(df['브랜드명'])
brand_list = []
mean_list = []
for key, group in df3:
    print("* key :", key)
    print("* number :", len(group))
    mean = round(group['별점'].mean(), 2)
    print(mean)
    mean_list.append(mean)
    brand_list.append(key)

rank = pd.Series(mean_list, index=brand_list)
print(rank.sort_values(ascending=False))

* key : 7번가피자
* number : 3263
4.83
* key : 김준현의피자헤븐
* number : 976
4.62
* key : 도미노피자
* number : 1186
4.6
* key : 미스터피자
* number : 1940
4.48
* key : 반올림피자샵
* number : 6096
4.75
* key : 파파존스피자
* number : 2007
4.65
* key : 피자나라치킨공주
* number : 2818
4.6
* key : 피자마루
* number : 472
4.66
* key : 피자알볼로
* number : 2555
4.71
* key : 피자헛
* number : 4325
4.58

7번가피자       4.83
반올림피자샵      4.75
피자알볼로       4.71
피자마루        4.66
파파존스피자      4.65
김준현의피자헤븐    4.62
피자나라치킨공주    4.60
도미노피자       4.60
피자헛         4.58
미스터피자       4.48
'''

'''
# 주문그룹별 주문건수 파악 및 주문그룹별 선호 메뉴구분 분석
df1 = df.groupby(df['그룹구분'])

for key, group in df1:
    print("="*50)
    print("* key :", key)
    print("="*50)
    vsum = 0
    for i in menu:
        count = 0
        for j in group['메뉴구분']:
            if i in j:
                count += 1
        vsum += count
        print(f"* {i}: {count}")
    print("-"*50)
    normalper = round((vsum/len(group))*100, 2)
    etcper = round(((len(group) - vsum)/len(group))*100, 2)
    print(">>> 일반 합계:", vsum, " ,비율:", normalper, "%")
    print(">>> 기타:", (len(group) - vsum), " ,비율:", etcper, "%")
    print(">>> 총 합계:", len(group), " ,비율: 100%")
    print('\n')
'''

'''
==================================================
* key : Party
==================================================
* 베이컨: 300
* 고르곤졸라: 20
* 콤비네이션: 435
* 스테이크: 294
* 쉬림프: 426
* 포테이토: 507
* 치킨: 453
* 고구마: 318
* 불고기: 410
* 하와이안: 105
* 페페로니: 144
--------------------------------------------------
>>> 일반 합계: 3412  ,비율: 37.13 %
>>> 기타: 5777  ,비율: 62.87 %
>>> 총 합계: 9189  ,비율: 100%


==================================================
* key : Side
==================================================
* 베이컨: 230
* 고르곤졸라: 19
* 콤비네이션: 521
* 스테이크: 549
* 쉬림프: 473
* 포테이토: 764
* 치킨: 642
* 고구마: 351
* 불고기: 368
* 하와이안: 50
* 페페로니: 185
--------------------------------------------------
>>> 일반 합계: 4152  ,비율: 60.36 %
>>> 기타: 2727  ,비율: 39.64 %
>>> 총 합계: 6879  ,비율: 100%


==================================================
* key : Single
==================================================
* 베이컨: 250
* 고르곤졸라: 39
* 콤비네이션: 1017
* 스테이크: 946
* 쉬림프: 911
* 포테이토: 1324
* 치킨: 922
* 고구마: 399
* 불고기: 461
* 하와이안: 102
* 페페로니: 214
--------------------------------------------------
>>> 일반 합계: 6585  ,비율: 68.81 %
>>> 기타: 2985  ,비율: 31.19 %
>>> 총 합계: 9570  ,비율: 100%
'''
'''
# 서울권/비서울권 브랜드 별점 합계 & 리뷰 건수 파악을 통한 브랜드 선호도 파악
df4 = df.groupby(df['브랜드명'])
for key, group in df4:
    print("="*50)
    print("* key :", key)
    print("="*50)
    seoul = len(group[group['지역구분'] == "서울특별시"])
    seoul_mean = round(group[group['지역구분'] == "서울특별시"]['별점'].mean(), 2)
    notseoul = len(group[group['지역구분'] != "서울특별시"])
    not_mean = round(group[group['지역구분'] != "서울특별시"]['별점'].mean(), 2)
    print(f">>> 서울지역: 평균 {seoul_mean}점 (리뷰 {seoul}개)")
    print(f">>> 비서울지역: 평균 {not_mean}점 (리뷰 {notseoul}개)")
    print("-"*50)
'''
'''
==================================================
* key : 7번가피자
==================================================
>>> 서울지역: 평균 4.82점 (리뷰 2072개)
>>> 비서울지역: 평균 4.85점 (리뷰 1191개)
--------------------------------------------------
==================================================
* key : 김준현의피자헤븐
==================================================
>>> 서울지역: 평균 4.63점 (리뷰 861개)
>>> 비서울지역: 평균 4.56점 (리뷰 115개)
--------------------------------------------------
==================================================
* key : 도미노피자
==================================================
>>> 서울지역: 평균 4.6점 (리뷰 559개)
>>> 비서울지역: 평균 4.6점 (리뷰 627개)
--------------------------------------------------
==================================================
* key : 미스터피자
==================================================
>>> 서울지역: 평균 4.47점 (리뷰 1167개)
>>> 비서울지역: 평균 4.49점 (리뷰 773개)
--------------------------------------------------
==================================================
* key : 반올림피자샵
==================================================
>>> 서울지역: 평균 4.74점 (리뷰 3577개)
>>> 비서울지역: 평균 4.76점 (리뷰 2519개)
--------------------------------------------------
==================================================
* key : 파파존스피자
==================================================
>>> 서울지역: 평균 4.65점 (리뷰 1349개)
>>> 비서울지역: 평균 4.66점 (리뷰 658개)
--------------------------------------------------
==================================================
* key : 피자나라치킨공주
==================================================
>>> 서울지역: 평균 4.6점 (리뷰 2144개)
>>> 비서울지역: 평균 4.6점 (리뷰 674개)
--------------------------------------------------
==================================================
* key : 피자마루
==================================================
>>> 서울지역: 평균 4.66점 (리뷰 380개)
>>> 비서울지역: 평균 4.64점 (리뷰 92개)
--------------------------------------------------
==================================================
* key : 피자알볼로
==================================================
>>> 서울지역: 평균 4.72점 (리뷰 1360개)
>>> 비서울지역: 평균 4.71점 (리뷰 1195개)
--------------------------------------------------
==================================================
* key : 피자헛
==================================================
>>> 서울지역: 평균 4.6점 (리뷰 2426개)
>>> 비서울지역: 평균 4.57점 (리뷰 1899개)
--------------------------------------------------
'''

'''
# 브랜드 메뉴구분별 별점 만족도 분석(1~2점 불만족, 3점 보통, 4~5점 만족)
df4 = df.groupby(df['브랜드명'])
for key, group in df4:
    print("="*50)
    print("* key :", key)
    print("="*50)
    
    for i in menu:
        for j in range(len(group)):
            if i in group['메뉴구분'][j]:
                try:
                    vsum = group[group[i in group['메뉴구분'][j]]]['별점'].mean()
                    print(f"$ MENU {i}: {vsum}")
                except:
                    pass
    print("-"*50)
'''

# [] 브랜드별 메뉴구분별 대표 메뉴 점수 분석 (값은 별점 합계로 계산)
df5 = df.groupby(['브랜드명', '메뉴구분'])
for key, group in df5:
    if len(group) > 500:
        print("="*50)
        print("* key :", key)
        print("* review :", format(len(group), ","))
        point = group['별점'].sum()
        print("* point : {}점".format(format(point, ",")))
        print("")

'''
==================================================
* key : ('7번가피자', '포테이토R')
* review : 712
* point : 3,490점

==================================================
* key : ('7번가피자', '홍대피자R')
* review : 567
* point : 2,750점

==================================================
* key : ('도미노피자', '포테이토피자')
* review : 504
* point : 2,386점

==================================================
* key : ('반올림피자샵', '반반L사이즈（3~4인）')
* review : 984
* point : 4,701점

==================================================
* key : ('반올림피자샵', '반반R사이즈（2~3인）')
* review : 546
* point : 2,599점

==================================================
* key : ('반올림피자샵', '세트1（피자L＋스파게티＋콜라500ml）')
* review : 1,511
* point : 7,241점

==================================================
* key : ('반올림피자샵', '세트5（피자R＋스파게티＋콜라500ml）')
* review : 532
* point : 2,543점

==================================================
* key : ('반올림피자샵', '오리지널피자L＋오리지널피자L')
* review : 663
* point : 3,150점

==================================================
* key : ('반올림피자샵', '포텐피자L')
* review : 591
* point : 2,825점

==================================================
* key : ('파파존스피자', '수퍼파파스오리지널크러스트L')
* review : 648
* point : 2,951점

==================================================
* key : ('파파존스피자', '수퍼파파스오리지널크러스트R')
* review : 657
* point : 3,003점

==================================================
* key : ('피자나라치킨공주', '피치세트L')
* review : 584
* point : 2,682점

==================================================
* key : ('피자나라치킨공주', '피치세트M')
* review : 546
* point : 2,489점

==================================================
* key : ('피자알볼로', '콤비네이션피자')
* review : 744
* point : 3,479점

==================================================
* key : ('피자헛', '메가더블세트')
* review : 824
* point : 3,688점
'''
print(df.describe()) # 별점/맛/양 기초통계량 파악/ 배달 점수는 제외
# [v] 메뉴구분별 별점 만족도 분석(평균 13.8점 최대 15점)
df5 = df.groupby(['브랜드명', '메뉴구분'])
for key, group in df5:
    if len(group) > 500:
        print("="*50)
        print("* key :", key)
        print("* review :", format(len(group), ","))
        spoint = group['별점'].mean()
        print(">>> star point : {}점".format(round(spoint, 2)))
        tpoint = group['맛'].mean()
        print(">>> taste point : {}점".format(round(tpoint, 2)))
        vpoint = group['별점'].mean()
        print(">>> volume point : {}점".format(round(vpoint, 2)))
        total = spoint + tpoint + vpoint
        if total > 14.4:
            com = "매우 높음"
        elif total > 13.8:
            com = "높음"
        elif total > 13.2:
            com = "보통"
        else:
            com = "낮음"
        print("-"*50)
        print(f"이 메뉴의 추천 정도는 '{com}'입니다.")
        print("")


'''
==================================================
* key : ('7번가피자', '포테이토R')
* review : 712
>>> star point : 4.9점
>>> taste point : 4.88점
>>> volume point : 4.9점
--------------------------------------------------
이 메뉴의 추천 정도는 '매우 높음'입니다.

==================================================
* key : ('7번가피자', '홍대피자R')
* review : 567
>>> star point : 4.85점
>>> taste point : 4.75점
>>> volume point : 4.85점
--------------------------------------------------
이 메뉴의 추천 정도는 '매우 높음'입니다.

==================================================
* key : ('도미노피자', '포테이토피자')
* review : 504
>>> star point : 4.73점
>>> taste point : 4.65점
>>> volume point : 4.73점
--------------------------------------------------
이 메뉴의 추천 정도는 '높음'입니다.

==================================================
* key : ('반올림피자샵', '반반L사이즈（3~4인）')
* review : 984
>>> star point : 4.78점
>>> taste point : 4.78점
>>> volume point : 4.78점
--------------------------------------------------
이 메뉴의 추천 정도는 '높음'입니다.

==================================================
* key : ('반올림피자샵', '반반R사이즈（2~3인）')
* review : 546
>>> star point : 4.76점
>>> taste point : 4.75점
>>> volume point : 4.76점
--------------------------------------------------
이 메뉴의 추천 정도는 '높음'입니다.

==================================================
* key : ('반올림피자샵', '세트1（피자L＋스파게티＋콜라500ml）')
* review : 1,511
>>> star point : 4.79점
>>> taste point : 4.76점
>>> volume point : 4.79점
--------------------------------------------------
이 메뉴의 추천 정도는 '높음'입니다.

==================================================
* key : ('반올림피자샵', '세트5（피자R＋스파게티＋콜라500ml）')
* review : 532
>>> star point : 4.78점
>>> taste point : 4.76점
>>> volume point : 4.78점
--------------------------------------------------
이 메뉴의 추천 정도는 '높음'입니다.

==================================================
* key : ('반올림피자샵', '오리지널피자L＋오리지널피자L')
* review : 663
>>> star point : 4.75점
>>> taste point : 4.66점
>>> volume point : 4.75점
--------------------------------------------------
이 메뉴의 추천 정도는 '높음'입니다.

==================================================
* key : ('반올림피자샵', '포텐피자L')
* review : 591
>>> star point : 4.78점
>>> taste point : 4.75점
>>> volume point : 4.78점
--------------------------------------------------
이 메뉴의 추천 정도는 '높음'입니다.

==================================================
* key : ('파파존스피자', '수퍼파파스오리지널크러스트L')
* review : 648
>>> star point : 4.55점
>>> taste point : 4.43점
>>> volume point : 4.55점
--------------------------------------------------
이 메뉴의 추천 정도는 '보통'입니다.

==================================================
* key : ('파파존스피자', '수퍼파파스오리지널크러스트R')
* review : 657
>>> star point : 4.57점
>>> taste point : 4.53점
>>> volume point : 4.57점
--------------------------------------------------
이 메뉴의 추천 정도는 '보통'입니다.

==================================================
* key : ('피자나라치킨공주', '피치세트L')
* review : 584
>>> star point : 4.59점
>>> taste point : 4.49점
>>> volume point : 4.59점
--------------------------------------------------
이 메뉴의 추천 정도는 '보통'입니다.

==================================================
* key : ('피자나라치킨공주', '피치세트M')
* review : 546
>>> star point : 4.56점
>>> taste point : 4.37점
>>> volume point : 4.56점
--------------------------------------------------
이 메뉴의 추천 정도는 '보통'입니다.

==================================================
* key : ('피자알볼로', '콤비네이션피자')
* review : 744
>>> star point : 4.68점
>>> taste point : 4.57점
>>> volume point : 4.68점
--------------------------------------------------
이 메뉴의 추천 정도는 '높음'입니다.

==================================================
* key : ('피자헛', '메가더블세트')
* review : 824
>>> star point : 4.48점
>>> taste point : 4.45점
>>> volume point : 4.48점
--------------------------------------------------
이 메뉴의 추천 정도는 '보통'입니다.
'''
